Fix UI string: remove placeholders for arguments
Summary:
%1, %2,... are interpreted as placeholders for arguments leading to
an I18n_ARGUMENT_MISSING error in the visible string. Rewrite them
as more generic placeholder which are not replaced, as it was done
for other strings in the same file.

Test Plan:
When accessing the configuration of 'Diff & Merge', "What's this"
on "External diff display", there is no more I18n_ARGUMENT_MISSING
error.

Reviewers: chehrlic, lueck

Reviewed By: chehrlic, lueck

Subscribers: lueck, #localization

Tags: #localization, #subversion_client_kde_svn

Differential Revision: https://phabricator.kde.org/D4356