make Targets::fromUrlList() return an url instead a path if requested
Summary:
Targets::fromUrlList() tries to be smart and returns a path if the url is a local file.
This is wrong when kdesvn is working on a local svn repository (via file:///path/to/svn/repo) because subversion needs a url then.
Therefore I added a 'bool bKeepUrl' to make sure to create a url if needed.
This does fix the problem when working with kdesvn but will not help if the request is done via kio_svn.
There we have to do a check if we're working on a wc or repo - don't know if it's worth the trouble ...

Test Plan: Open kdesvn on a local svn repo, try to move/copy a file - works now (before there was a error message)

Reviewers: dfaure

Reviewed By: dfaure

Subscribers: broulik

Differential Revision: https://phabricator.kde.org/D7609