Summary: Properly emit dataChanged signals when a file changes. dataChanged() was only emitted sometimes and then only for column 0.
Test Plan:
Open kdesvn, externally modify a file or do a svn revert. The item immediately
changes it's color. Before the change the color only changed when the item
was redrawn (e.g. when the focus changed).
It still is not updated correctly in the dirview. Needs some more investigation.

Reviewers: dfaure, aspotashev, broulik

Differential Revision: https://phabricator.kde.org/D11337