Make sqlite test pass when running in any timezone
SVN revisions are sensitive to the timezone, so specify the exact time
and timezone for the test to pass.

E.g. the revision {2014-09-27} in UTC resolves to 1400903, as it was
made at Fri Sep 26 23:16:54 2014 UTC.

The same revision {2014-09-27} in BST (UTC +0001) resolves to 1400900,
as it was made at Fri Sep 26 22:17:34 2014 UTC, i.e. 23:17:34 in BST
time.